util.c:153:11: error: implicit conversion loses integer precision: 'u_long' (aka 'unsigned long') to 'uint32_t' (aka 'unsigned int') [-Werror,-Wshorten-64-to-32]

+++ test/format/util.c
@@ -150,7 +150,7 @@ fp_readv(FILE *fp, char *name, bool eof_ok, uint32_t *vp)
     ulv = strtoul(buf, &endptr, 10);
     testutil_assert(errno == 0 && endptr[0] == '\n');
     testutil_assert(ulv <= UINT32_MAX);
-    *vp = ulv;
+    *vp = (uint32_t)ulv;
     return (false);
 }
